Separate file at specific path
